refactor: move AbstractRepositoryPersistenceBean to JakartaPersistenceExtension and update bean inheritance
