+++ jnosql-jakarta-persistence/jnosql-jakarta-persistence-driver/src/main/java/org/eclipse/jnosql/extensions/sql/repository/spi/RepositoryPersistenceBean.java
@@ -13,11 +13,12 @@
  *   Otavio Santana
  *   Ondro Mihalyi
  */
-package org.eclipse.jnosql.jakartapersistence.mapping.repository;
+package org.eclipse.jnosql.extensions.sql.repository.spi;
 
 import jakarta.data.repository.DataRepository;
 import jakarta.enterprise.inject.spi.BeanManager;
 import org.eclipse.jnosql.jakartapersistence.mapping.PersistenceDocumentTemplate;
+import org.eclipse.jnosql.jakartapersistence.mapping.repository.JakartaPersistenceRepositoryProxy;
 import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.core.spi.AbstractBean;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
@@ -37,7 +38,7 @@ import java.lang.reflect.InvocationHandler;
  * @param <T> the type of the repository
  * @see AbstractBean
  */
-public class RepositoryPersistenceBean<T extends DataRepository<T, ?>> extends AbstractRepositoryPersistenceBean<T> {
+public class RepositoryPersistenceBean<T extends DataRepository<T, ?>> extends JakartaPersistenceExtension.AbstractRepositoryPersistenceBean<T> {
 
     /**
      * Constructor
